backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Document processing
python-docx==1.1.0
PyPDF2==3.0.1
python-magic-bin==0.4.14; platform_system == "Windows"

# AI/ML
google-generativeai==0.3.2
exa-py==1.0.9
numpy==1.26.3
scikit-learn==1.4.0

# Utilities
aiofiles==23.2.1
httpx==0.26.0
pyyaml==6.0.1

# UOIONHHC - Dependencies
